refactor: add send/ directory
