+++ src/components/ChatInput.tsx
@@ -1,5 +1,5 @@
 import type { Component } from 'solid-js'
-import { sendChatInfoToAll } from '../utils/sendChatInfo'
+import { sendChatInfoToAll } from '../utils/send/sendChatInfo'
 import { setChatInfos, localChatInfoFrom } from '../utils/chat'
 
 const ChatInput: Component<{ chatGroup: number }> = (props) => {
