+++ src/utils/send/sendChatInfo.ts
@@ -1,5 +1,5 @@
-import { RemoteUserInfo } from './user'
-import { ChatInfo, localChatInfoFrom, ChatBoxInfo, getChatInfos } from './chat'
+import { RemoteUserInfo } from '../user'
+import { ChatInfo, localChatInfoFrom, ChatBoxInfo, getChatInfos } from '../chat'
 import { sendTo, sendToAll } from './send'
 
 export const sendChatInfoToAll = (data: ChatInfo) => {
